pyproject.toml
# ---------------------------------------------------------------------------
# tikzgif -- Parameterized TikZ to animated GIF pipeline
# ---------------------------------------------------------------------------
# Build backend: hatchling
# ---------------------------------------------------------------------------

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

# ---------------------------------------------------------------------------
# Project metadata (PEP 621)
# ---------------------------------------------------------------------------
[project]
name = "tikzgif"
version = "0.1.0"
description = "Convert parameterized TikZ/LaTeX files into animated GIFs."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "J.C. Vaught", email = "jvaught@sc.edu" },
]
keywords = ["tikz", "latex", "gif", "animation", "visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Text Processing :: Markup :: LaTeX",
    "Typing :: Typed",
]

# --- Core runtime dependencies ------------------------------------------------
dependencies = [
    "Pillow>=10.0.0",
    "tqdm>=4.60.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "PyYAML>=6.0.0",
]

# --- Optional dependency groups -----------------------------------------------
[project.optional-dependencies]
pdf2svg = [
    "cairosvg>=2.7.0",
]
watch = [
    "watchdog>=3.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "mypy>=1.5.0",
    "ruff>=0.4.0",
    "pre-commit>=3.4.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-click>=5.0.0",
]
all = [
    "tikzgif[pdf2svg,watch,dev,docs]",
]

# --- URLs ---------------------------------------------------------------------
[project.urls]
Homepage = "https://github.com/j-vaught/tikzgif"
Documentation = "https://tikzgif.readthedocs.io"
Repository = "https://github.com/j-vaught/tikzgif"
Issues = "https://github.com/j-vaught/tikzgif/issues"
Changelog = "https://github.com/j-vaught/tikzgif/blob/main/CHANGELOG.md"

# --- CLI entry points ---------------------------------------------------------
[project.scripts]
tikzgif = "tikzgif.cli.main:cli_entry"

# ---------------------------------------------------------------------------
# Hatchling build configuration (src layout)
# ---------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["tikzgif"]

[tool.hatch.build.targets.sdist]
include = [
    "tikzgif/**",
    "tests/**",
    "LICENSE",
    "README.md",
    "CHANGELOG.md",
    "pyproject.toml",
]

# ---------------------------------------------------------------------------
# Tool configurations
# ---------------------------------------------------------------------------

# --- Ruff (linting + formatting) -------------------------------------------
[tool.ruff]
target-version = "py310"
line-length = 88
src = ["tikzgif", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["tikzgif"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- Mypy -------------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
packages = ["tikzgif"]

[[tool.mypy.overrides]]
module = [
    "cairosvg.*",
    "PIL.*",
    "tqdm.*",
]
ignore_missing_imports = true

# --- Pytest ------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "-v",
]
markers = [
    "integration: marks tests requiring a LaTeX installation (deselect with '-m \"not integration\"')",
    "slow: marks tests that take more than 10 seconds",
    "benchmark: performance benchmark test",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:tqdm",
    "ignore::ResourceWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]

# --- Coverage ----------------------------------------------------------------
[tool.coverage.run]
source = ["tikzgif"]
branch = true

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "@overload",
    "raise NotImplementedError",
]
